+++ lib/rubocop/cop/style/space_around_operators.rb
@@ -40,8 +40,8 @@ module RuboCop
           if node.loc.operator # aref assignment, attribute assignment
             on_special_asgn(node)
           elsif !node.unary_operation? && !called_with_dot?(node)
-            op = node.loc.selector
-            if operator?(op.source.to_sym)
+            op = node.method_name
+            if op != :[] && op != :! && operator?(op)
               _, _, right, = *node
               check_operator(node.loc.selector, right.source_range)
             end